Add OrderPart Component so each part within an order can have its own price assigned
